+++ packages/chat_repository/lib/src/firebase_chat_repository.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:developer';
 import 'dart:io';
 
@@ -6,7 +7,9 @@ import 'package:chat_repository/src/chat_repo.dart';
 import 'package:chat_repository/src/models/chat.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_storage/firebase_storage.dart';
+import 'package:http/http.dart';
 import 'package:intl/intl.dart';
+import 'package:user_repository/src/models/my_user.dart';
 import 'package:uuid/uuid.dart';
 
 class FirebaseChatRepository implements ChatRepository {
@@ -30,17 +33,18 @@ class FirebaseChatRepository implements ChatRepository {
   }
 
   @override
-  Future<void> deleteChat(String chatID, DateTime send) async {
+  Future<void> deleteChat(Chat chat, DateTime send) async {
     try {
       String formattedTime = DateFormat('yyyyMMddHHmmssSSS').format(send);
-
-      Reference firebaseStoreRef = FirebaseStorage.instance
-          .ref()
-          .child("ChatImage/${chatID}/${formattedTime}_lead");
-      await firebaseStoreRef.delete();
+      if(chat.type == "image") {
+        Reference firebaseStoreRef = FirebaseStorage.instance
+            .ref()
+            .child("ChatImage/${chat.chatID}/${formattedTime}_lead");
+        await firebaseStoreRef.delete();
+      }
 
       await chatCollection
-          .doc(chatID)
+          .doc(chat.chatID)
           .collection("message")
           .doc(formattedTime)
           .delete();
@@ -86,6 +90,45 @@ class FirebaseChatRepository implements ChatRepository {
           .collection('message')
           .doc(formattedTime)
           .set(chat.toEntity().toDocument());
+    } catch (e) {
+      log(e.toString());
+      rethrow;
+    }
+  }
+
+  @override
+  Future<void> sendNotificationChat( String token , String message , String name,String type) async {
+
+    try {
+      final url = 'https://fcm.googleapis.com/fcm/send';
+      final headers = {
+        'Authorization': 'key=AAAAANeFT70:APA91bFMcXywgTpYWakCnVvH6uxLHoZrcX-ze1OuGAOVEKFXWVbRDzjRBSIouEzTSo3dUw_E211jQThb_eMGY-dCzfZ-HoAXPm2yLqXnPLbyKql9XfF-rtWmYx35qpsHMUoGiMmBvLqS',
+        'Content-Type': 'application/json',
+      };
+      log("đã gửi notification");
+      if(type == 'image'){
+        final body = {
+          "to": token,
+          "notification": {
+            "title" : "Tin nhắn mới từ $name",
+            "body" : "$name đã gửi 1 ảnh cho bạn"
+          }
+
+        };
+        await post(Uri.parse(url), headers: headers , body: jsonEncode(body));
+      } else {
+        final body = {
+          "to": token,
+          "notification": {
+            "title" : "Tin nhắn mới từ $name",
+            "body" : message
+          }
+
+        };
+        await post(Uri.parse(url), headers: headers , body: jsonEncode(body));
+
+      }
+
     } catch (e) {
       log(e.toString());
       rethrow;
